commit after presentation

+++ src/AppBundle/Controller/ParticipantController.php
@@ -7,12 +7,16 @@ use Cocur\Slugify\Slugify;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Form\Extension\Core\Type\PasswordType;
+use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Intervention\Image\ImageManager;
+use Symfony\Component\Validator\Constraints\Length;
 
 /**
  * Participant controller.
@@ -30,9 +34,7 @@ class ParticipantController extends Controller
     public function indexAction()
     {
         $em = $this->getDoctrine()->getManager();
-
         $participants = $em->getRepository('AppBundle:Participant')->findAll();
-
         return $this->render('participant/index.html.twig', array(
             'participants' => $participants,
         ));
@@ -51,50 +53,54 @@ class ParticipantController extends Controller
      */
     public function InscriptionAction(Request $request, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder)
     {
-        $participant = new Participant();
+        if (!$this->getUser() or $this->isGranted('ROLE_ADMIN')){
+            $participant = new Participant();
 
-        $participant->setAdministrateur(0);
-        $participant->setActif(1);
+            $participant->setAdministrateur(0);
+            $participant->setActif(1);
 
-        $formInscription = $this->createForm('AppBundle\Form\ParticipantType', $participant);
+            $formInscription = $this->createForm('AppBundle\Form\ParticipantType', $participant);
 
-        $formInscription->remove('administrateur');
-        $formInscription->remove('actif');
+            $formInscription->remove('administrateur');
+            $formInscription->remove('actif');
 
-        $formInscription->handleRequest($request);
+            $formInscription->handleRequest($request);
 
-        if ($formInscription->isValid() && $formInscription->isSubmitted()) {
-            $participant->setRoles(['ROLE_USER']);
-            $participant->setSalt('IlPleutIlMouille');
-            $passwordSale = $passwordEncoder->encodePassword($participant, $participant->getMotDePasseParticipant());
-            $participant->setMotDePasseParticipant($passwordSale);
-            dump($participant->getPathImage());
-            die();
-            if ($participant->getPathImage() == null){
-                $participant->setPathImage($this->get('kernel')->getProjectDir()."\web\images\photoProfil\Avatar_vide.png");
-            }else{
+            if ($formInscription->isValid() && $formInscription->isSubmitted()) {
+                $participant->setRoles(['ROLE_USER']);
+                $participant->setSalt('IlPleutIlMouille');
+                $passwordSale = $passwordEncoder->encodePassword($participant, $participant->getMotDePasseParticipant());
+                $participant->setMotDePasseParticipant($passwordSale);
+
                 //Traitement de l'image
                 /**
-                 * @var UploadedFile $image
+                 * @var UploadedFile $uploadedFile
                  */
-                $slugify = new Slugify();
-                $filename = $slugify->slugify($participant->getPseudo()) . "-" . $participant->getId() . ".jpg";
 
-                $image = $participant->getPathImage();
+                if($uploadedFile = $participant->getPathImage()){
 
-                $imageManager = new ImageManager();
-                $imageOrigine = $imageManager->make($image);
-                $imageOrigine->resize(120, 150);
+                    //upload image
+                    $imageManager = new ImageManager();
 
-                $imageOrigine->save($this->get('kernel')->getProjectDir()."\web\images\photoProfil\\".$filename);
-                $participant->setPathImage('/images/photoProfil/' . $filename);
-            }
-            
-            $em->persist($participant);
-            $em->flush();
+                    $imageOrigine = $imageManager->make($uploadedFile);
+                    $imageOrigine->resize(120, 150);
 
-            $this->addFlash('success', 'Le participant a bien été enregistré');
-            return $this->redirectToRoute('participant_afficherProfil', array('id' => $participant->getId()));
+                    $slugify = new Slugify();
+                    $filename = $slugify->slugify($participant->getPseudo()) . "-" . $participant->getId() . ".jpg";
+
+                    $imageOrigine->save($this->get('kernel')->getProjectDir()."\web\images\photoProfil\\".$filename);
+                    $participant->setPathImage('/images/photoProfil/' . $filename);
+                }
+                else
+                {
+                    // image par default
+                    $participant->setPathImage('/images/photoAvatar/avatar.jpg');//NE PAS METTRE "WEB" DANS LE CHEMIN D'ACCES!
+                }
+                $em->persist($participant);
+                $em->flush();
+
+                $this->addFlash('success', 'Le participant a bien été enregistré');
+                return $this->redirectToRoute('homepage');
         }
 
         return $this->render('participant/inscription.html.twig', array(
@@ -102,6 +108,10 @@ class ParticipantController extends Controller
             'form' => $formInscription->createView(),
         ));
 
+        } else{
+            $this->addFlash('error', 'Création de compte impossible si vous êtes connecté !');
+            return $this->redirectToRoute('homepage');
+        }
     }
 
     /**
@@ -120,7 +130,6 @@ class ParticipantController extends Controller
         ));
     }
 
-
     /**
      * EDITER LE PROFIL D'UN PARTICIPANT
      * @Route("/edit/{id}", name="participant_edit")
@@ -132,56 +141,114 @@ class ParticipantController extends Controller
      * @param UserPasswordEncoderInterface $passwordEncoder
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
-
     public function editAction(Request $request, Participant $participant, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder)
     {
-        if ($this->getUser() == $participant or $this->isGranted("ROLE_ADMIN")){
-            $deleteForm= $this->createDeleteForm($participant);
+        if ($this->getUser() == $participant or $this->isGranted("ROLE_ADMIN")) {
+            $deleteForm = $this->createDeleteForm($participant);
+            $editForm = $this->createForm('AppBundle\Form\ParticipantType', $participant);
+            $motDePasse = $participant->getMotDePasseParticipant();
 
-            $editForm= $this->createForm('AppBundle\Form\ParticipantType', $participant);
+            $oldPhotoProfile = $participant->getPathImage();
+            $editForm->remove('motDePasseParticipant');
             $editForm->handleRequest($request);
 
-            if ($editForm->isSubmitted()&& $editForm->isValid()) {
-
-                $passwordSale = $passwordEncoder->encodePassword($participant, $participant->getMotDePasseParticipant());
-                $participant->setMotDePasseParticipant($passwordSale);
-
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $participant->setMotDePasseParticipant($motDePasse);
+//                    $passwordSale = $passwordEncoder->encodePassword($participant, $participant->getMotDePasseParticipant());
+//                    $participant->setMotDePasseParticipant($passwordSale);
 
                 //Traitement de l'image
                 /**
                  * @var UploadedFile $image
                  */
-                $slugify = new Slugify();
-                $filename = $slugify->slugify($participant->getPseudo()) . "-" . $participant->getId() . ".jpg";
-
-                $image = $participant->getPathImage();
-
-                $imageManager = new ImageManager();
-                $imageOrigine = $imageManager->make($image);
-                $imageOrigine->resize(120, 150);
-
-                $imageOrigine->save($this->get('kernel')->getProjectDir()."\web\images\photoProfil\\".$filename);
-                $participant->setPathImage('/images/photoProfil/' . $filename);
-
+                if ($participant->getPathImage() == null) {
+                    $participant->setPathImage($oldPhotoProfile);
+                } else {
+                    //Traitement de l'image
+                    /**
+                     * @var UploadedFile $image
+                     */
+                    $slugify = new Slugify();
+                    $filename = $slugify->slugify($participant->getPseudo()) . "-" . $participant->getId() . ".jpg";
+                    $image = $participant->getPathImage();
+                    $imageManager = new ImageManager();
+                    $imageOrigine = $imageManager->make($image);
+                    $imageOrigine->resize(120, 150);
+                    $imageOrigine->save($this->get('kernel')->getProjectDir() . "\web\images\photoProfil\\" . $filename);
+                    $participant->setPathImage('/images/photoProfil/' . $filename);
+                }
                 $em->persist($participant);
                 $em->flush();
-
                 $this->addFlash('success', 'Le participant a bien été modifié');
-
-                return $this->redirectToRoute('participant_afficherProfil', array('id' =>$participant->getId()));
+                return $this->redirectToRoute('participant_afficherProfil', array('id' => $participant->getId()));
             }
             return $this->render('participant/edit.html.twig', array(
-                'participant' =>$participant,
+                'participant' => $participant,
                 'edit_form' => $editForm->createView(),
                 'delete_form' => $deleteForm->createView(),
             ));
-        }else{
+        } else {
             $this->addFlash('error', 'La modification d\'un utilisateur autre que vous-même est IMPOSSIBLE');
             return $this->redirectToRoute('homepage');
         }
+    }
 
+    /**
+     * @param Request $request
+     * @param Participant $participant
+     * @param EntityManagerInterface $em
+     * @param UserPasswordEncoderInterface $passwordEncoder
+     * @Route("/edit/{id}/password", name="participant_edit_password")
+     */
+    public function editPasswordAction(Request $request, Participant $participant, EntityManagerInterface $em, UserPasswordEncoderInterface $passwordEncoder)
+    {
+        if ($this->getUser() == $participant or $this->isGranted("ROLE_ADMIN")) {
+            $formBuilder = $this->createFormBuilder();
+            $formBuilder->add('motDePasseActuel', PasswordType::class, [
+                'label' => 'Ancien mot de passe'])
+                ->add('motDePasseParticipant', RepeatedType::class, [
+                    'type' => PasswordType::class,
+                    'invalid_message' => 'Le mot de passe doit être identique',
+                    'options' => ['attr' => ['class' => 'password-field']],
+                    'required' => true,
+                    'first_options' => ['label' => 'Nouveau Mot de passe',
+                        'constraints' => new Length(['min' => 5]),
+                        'invalid_message'=> 'Le mot de passe doit contenir minimum  5 caractères et/ou chiffres'],
+                    'second_options' => ['label' => 'Confirmation nouveau mot de passe',
+                        'constraints' => new Length(['min' => 5]),],
+                ])
+                ->add('Enregistrer', SubmitType::class);
+            $form = $formBuilder->getForm();
+            $form->handleRequest($request);
 
-   }
+            if ($form->isSubmitted() && $form->isValid()) {
+                //salage du mot de passe saisi pour savoir si il correspond a celui en base de donnee
+                $passwordSaisi = $form->get('motDePasseActuel')->getData();
+                $encoderService = $this->container->get('security.password_encoder');
+                $match = $encoderService->isPasswordValid($participant, $passwordSaisi);
+                if ($match) {
+                    $participant->setSalt('IlPleutIlMouille');
+                    $passwordSale = $passwordEncoder->encodePassword($participant, $form->get('motDePasseParticipant')->getData());
+                    $participant->setMotDePasseParticipant($passwordSale);
+                    $em->persist($participant);
+                    $em->flush();
+                    $this->addFlash('success', 'Votre mot de passe a bien été modifié !');
+                    return $this->redirectToRoute('participant_afficherProfil', array('id' => $participant->getId()));
+                } else {
+                    $this->addFlash('error', 'L\'ancien mot de passe saisi est incorrecte');
+                    return $this->redirectToRoute('participant_edit_password', array('id' => $participant->getId()));
+                }
+            }
+            return $this->render('participant/editMotDePasse.html.twig', array(
+                'participant' => $participant,
+                'form' => $form->createView(),
+            ));
+        } else {
+            $this->addFlash('error', 'La modification d\'un utilisateur autre que vous-même est IMPOSSIBLE');
+            return $this->redirectToRoute('homepage');
+        }
+    }
+
 
 
     /**
@@ -204,25 +271,28 @@ class ParticipantController extends Controller
 
     public function deleteAction(Request $request, Participant $participant, EntityManagerInterface $em)
     {
-       $formBuilder = $this->createFormBuilder();
-       $formBuilder->add('Supprimer le participant', SubmitType::class);
+        if ($this->isGranted("ROLE_ADMIN")) {
+            $formBuilder = $this->createFormBuilder();
+            $formBuilder->add('Supprimer le participant', SubmitType::class);
 
-       $form = $formBuilder->getForm();
+            $form = $formBuilder->getForm();
 
-       $form->handleRequest($request);
-       if ($form->isSubmitted()) {
-           $em->remove($participant);
-           $em->flush();
+            $form->handleRequest($request);
+            if ($form->isSubmitted()) {
+                $em->remove($participant);
+                $em->flush();
 
-           $this->addFlash('success', 'Le participant a bien été supprimé');
-           return $this->redirectToRoute('participant_index');
+                $this->addFlash('success', 'Le participant a bien été supprimé');
+                return $this->redirectToRoute('participant_index');
+            }
+
+            return $this->render('participant/delete.html.twig', [
+                'participants' => $participant,
+                'form' => $form->createView()]);
+       }else{
+                $this->addFlash('error', 'La suppression d\'un utilisateur autre que l\'administrateur est IMPOSSIBLE');
+                return $this->redirectToRoute('homepage');
        }
-
-       return $this->render('participant/delete.html.twig', [
-        'participants' =>$participant,
-        'form' => $form->createView()
-    ]);
-
     }
